Logically separate views.
The single views file was becoming unwieldy, so they've been split up by
the section of the app they're responsible for. A new module containing
helper functions for the views was also added, to keep the view files
clean.
