+++ qmlist/views/admin_console/__init__.py
@@ -1,0 +1,10 @@
+from flask_security import login_required, roles_required
+
+from qmlist.qmlist import app
+from qmlist.views.admin_console import categories, items, shopping_lists, users
+
+@app.route("/admin")
+@login_required
+@roles_required("admin")
+def admin_console_home():
+    pass
